requirements.txt
pyautogui
customtkinter
pillow
paddleocr